generate_token() will raise if user is None
(change) `.generate_token()` will raise NotAuthorizedError if user is None.
Until now this had to be implemented anyway in the subclasses in
`.user_from_payload()`. This was a bit confusing, makes more sense
to put this code here as you can’t generate token if you don’t have
a user to generate the token for. We shouldn’t wait for a fail in
`.user_from_payload()`.